2.5.0

+++ LiteLoader.NET/Kernel/FormUI.cpp
@@ -2,6 +2,8 @@
 
 #include "../Header/MC/Player.hpp"
 
+#include "../Header/Logger/Logger.hpp"
+
 
 namespace LLNET::Form
 {
@@ -312,7 +314,7 @@ namespace LLNET::Form
     }
 
 
-    inline Slider::Slider(String^ name, String^ title, int min, int max, int step, int def)
+    inline Slider::Slider(String^ name, String^ title, double min, double max, double step, double def)
     {
         Name = name;
         this->title = title;
@@ -322,12 +324,12 @@ namespace LLNET::Form
         this->def = def;
     }
 
-    inline Slider::Slider(String^ name, String^ title, int min, int max, int step)
+    inline Slider::Slider(String^ name, String^ title, double min, double max, double step)
         : Slider(name, title, min, max, step, 0)
     {
     }
 
-    inline Slider::Slider(String^ name, String^ title, int min, int max)
+    inline Slider::Slider(String^ name, String^ title, double min, double max)
         : Slider(name, title, min, max, 1, 0)
     {
     }
